requirements.txt
python-telegram-bot==20.7
requests
python-dotenv
amadeus==1.1.2